Cleaned up commenting. Separated layout delegate
