+++ Source/Common/IndexPath+SwiftGridView.swift
@@ -1,4 +1,4 @@
-// SwiftGridView.swift
+// IndexPath+SwiftGridView.swift
 // Copyright (c) 2016 Nathan Lampi (http://nathanlampi.com/)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
