+++ Source/SwiftGridViewDataSource.swift
@@ -1,4 +1,4 @@
-// SwiftGridView.swift
+// SwiftGridViewDataSource.swift
 // Copyright (c) 2016 Nathan Lampi (http://nathanlampi.com/)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
